Refactored common code from Salesforce*Activity classes to SalesforceActivityDelegate NB: hybrid and react native activity classes not changed (yet)
