+++ libs/SalesforceSDK/src/com/salesforce/androidsdk/ui/SalesforceExpandableListActivity.java
@@ -27,159 +27,58 @@
 package com.salesforce.androidsdk.ui;
 
 import android.app.ExpandableListActivity;
-import android.content.IntentFilter;
 import android.os.Bundle;
-
-import com.salesforce.androidsdk.accounts.UserAccountManager;
-import com.salesforce.androidsdk.app.SalesforceSDKManager;
-import com.salesforce.androidsdk.rest.ClientManager;
-import com.salesforce.androidsdk.rest.ClientManager.LoginOptions;
-import com.salesforce.androidsdk.rest.ClientManager.RestClientCallback;
-import com.salesforce.androidsdk.rest.RestClient;
-import com.salesforce.androidsdk.security.PasscodeManager;
-import com.salesforce.androidsdk.util.EventsObservable;
-import com.salesforce.androidsdk.util.EventsObservable.EventType;
-import com.salesforce.androidsdk.util.LogoutCompleteReceiver;
-import com.salesforce.androidsdk.util.UserSwitchReceiver;
+import android.view.KeyEvent;
 
 /**
  * Abstract base class for all Salesforce expandable list activities.
  *
  * @author bhariharan
  */
-public abstract class SalesforceExpandableListActivity extends ExpandableListActivity {
+public abstract class SalesforceExpandableListActivity extends ExpandableListActivity implements SalesforceActivityInterface {
 
-	private PasscodeManager passcodeManager;
-    private UserSwitchReceiver userSwitchReceiver;
-    private LogoutCompleteReceiver logoutCompleteReceiver;
+	private final SalesforceActivityDelegate delegate;
 
-	/**
-	 * Method that is called after the activity resumes once we have a RestClient.
-	 *
-	 * @param client RestClient instance.
-	 */
-	public abstract void onResume(RestClient client);
+	public SalesforceExpandableListActivity() {
+		super();
+		delegate = new SalesforceActivityDelegate(this);
+	}
 
 	@Override
 	protected void onCreate(Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
-
-		// Gets an instance of the passcode manager.
-		passcodeManager = SalesforceSDKManager.getInstance().getPasscodeManager();
-        userSwitchReceiver = new ActivityUserSwitchReceiver();
-        registerReceiver(userSwitchReceiver, new IntentFilter(UserAccountManager.USER_SWITCH_INTENT_ACTION));
-        logoutCompleteReceiver = new ActivityLogoutCompleteReceiver();
-        registerReceiver(logoutCompleteReceiver, new IntentFilter(SalesforceSDKManager.LOGOUT_COMPLETE_INTENT_ACTION));
-
-		// Lets observers know that activity creation is complete.
-		EventsObservable.get().notifyEvent(EventType.MainActivityCreateComplete, this);
+		delegate.onCreate();
 	}
 
-	@Override 
+	@Override
 	public void onResume() {
 		super.onResume();
-
-		// Brings up the passcode screen if needed.
-		if (passcodeManager.onResume(this)) {
-
-			// Gets login options.
-			final String accountType = SalesforceSDKManager.getInstance().getAccountType();
-	    	final LoginOptions loginOptions = SalesforceSDKManager.getInstance().getLoginOptions();
-
-			// Gets a rest client.
-			new ClientManager(this, accountType, loginOptions, SalesforceSDKManager.getInstance().shouldLogoutWhenTokenRevoked()).getRestClient(this, new RestClientCallback() {
-
-				@Override
-				public void authenticatedRestClient(RestClient client) {
-					if (client == null) {
-						SalesforceSDKManager.getInstance().logout(SalesforceExpandableListActivity.this);
-						return;
-					}
-					onResume(client);
-
-					// Lets observers know that rendition is complete.
-					EventsObservable.get().notifyEvent(EventType.RenditionComplete);
-				}
-			});
-		}
+		delegate.onResume();
 	}
 
 	@Override
 	public void onUserInteraction() {
-		passcodeManager.recordUserInteraction();
+		delegate.onUserInteraction();
 	}
 
-    @Override
-    public void onPause() {
-        super.onPause();
-    	passcodeManager.onPause(this);
-    }
-
-    @Override
-    public void onDestroy() {
-    	unregisterReceiver(userSwitchReceiver);
-        unregisterReceiver(logoutCompleteReceiver);
-    	super.onDestroy();
-    }
-
-    /**
-     * Refreshes the client if the user has been switched.
-     */
-	protected void refreshIfUserSwitched() {
-		if (passcodeManager.onResume(this)) {
-
-			// Gets login options.
-			final String accountType = SalesforceSDKManager.getInstance().getAccountType();
-	    	final LoginOptions loginOptions = SalesforceSDKManager.getInstance().getLoginOptions();
-
-			// Gets a rest client.
-			new ClientManager(this, accountType, loginOptions,
-					SalesforceSDKManager.getInstance().shouldLogoutWhenTokenRevoked()).getRestClient(this, new RestClientCallback() {
-
-				@Override
-				public void authenticatedRestClient(RestClient client) {
-					if (client == null) {
-						SalesforceSDKManager.getInstance().logout(SalesforceExpandableListActivity.this);
-						return;
-					}
-					onResume(client);
-
-					// Lets observers know that rendition is complete.
-					EventsObservable.get().notifyEvent(EventType.RenditionComplete);
-				}
-			});
-		}
+	@Override
+	public void onPause() {
+		super.onPause();
+		delegate.onPause();
 	}
 
-    /**
-     * Performs actions on logout complete.
-     */
-    protected void logoutCompleteActions() {
-    }
+	@Override
+	public void onDestroy() {
+		delegate.onDestroy();
+		super.onDestroy();
+	}
 
-    /**
-     * Acts on the user switch event.
-     *
-     * @author bhariharan
-     */
-    private class ActivityUserSwitchReceiver extends UserSwitchReceiver {
+	@Override
+	public boolean onKeyUp(int keyCode, KeyEvent event) {
+		return delegate.onKeyUp(keyCode, event);
+	}
 
-		@Override
-		protected void onUserSwitch() {
-			refreshIfUserSwitched();
-		}
-    }
-
-	/**
-	 * Acts on the logout complete event.
-	 *
-	 * @author bhariharan
-	 */
-	private class ActivityLogoutCompleteReceiver extends LogoutCompleteReceiver {
-
-		@Override
-		protected void onLogoutComplete() {
-            logoutCompleteActions();
-		}
+	@Override
+	public void logoutCompleteActions() {
 	}
 }
