+++ libs/SalesforceSDK/src/com/salesforce/androidsdk/ui/SalesforceListActivity.java
@@ -48,150 +48,50 @@ import com.salesforce.androidsdk.util.UserSwitchReceiver;
  *
  * @author bhariharan
  */
-public abstract class SalesforceListActivity extends ListActivity {
+public abstract class SalesforceListActivity extends ListActivity implements SalesforceActivityInterface {
 
-	private PasscodeManager passcodeManager;
-    private UserSwitchReceiver userSwitchReceiver;
-    private LogoutCompleteReceiver logoutCompleteReceiver;
+	private final SalesforceActivityDelegate delegate;
 
-	/**
-	 * Method that is called after the activity resumes once we have a RestClient.
-	 *
-	 * @param client RestClient instance.
-	 */
-	public abstract void onResume(RestClient client);
+	public SalesforceListActivity() {
+		super();
+		delegate = new SalesforceActivityDelegate(this);
+	}
 
 	@Override
 	protected void onCreate(Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
-
-		// Gets an instance of the passcode manager.
-		passcodeManager = SalesforceSDKManager.getInstance().getPasscodeManager();
-        userSwitchReceiver = new ActivityUserSwitchReceiver();
-        registerReceiver(userSwitchReceiver, new IntentFilter(UserAccountManager.USER_SWITCH_INTENT_ACTION));
-        logoutCompleteReceiver = new ActivityLogoutCompleteReceiver();
-        registerReceiver(logoutCompleteReceiver, new IntentFilter(SalesforceSDKManager.LOGOUT_COMPLETE_INTENT_ACTION));
-
-		// Lets observers know that activity creation is complete.
-		EventsObservable.get().notifyEvent(EventType.MainActivityCreateComplete, this);
+		delegate.onCreate();
 	}
 
-	@Override 
+	@Override
 	public void onResume() {
 		super.onResume();
-
-		// Brings up the passcode screen if needed.
-		if (passcodeManager.onResume(this)) {
-
-			// Gets login options.
-			final String accountType = SalesforceSDKManager.getInstance().getAccountType();
-	    	final LoginOptions loginOptions = SalesforceSDKManager.getInstance().getLoginOptions();
-
-			// Gets a rest client.
-			new ClientManager(this, accountType, loginOptions, SalesforceSDKManager.getInstance().shouldLogoutWhenTokenRevoked()).getRestClient(this, new RestClientCallback() {
-
-				@Override
-				public void authenticatedRestClient(RestClient client) {
-					if (client == null) {
-						SalesforceSDKManager.getInstance().logout(SalesforceListActivity.this);
-						return;
-					}
-					onResume(client);
-
-					// Lets observers know that rendition is complete.
-					EventsObservable.get().notifyEvent(EventType.RenditionComplete);
-				}
-			});
-		}
+		delegate.onResume();
 	}
 
 	@Override
 	public void onUserInteraction() {
-		passcodeManager.recordUserInteraction();
+		delegate.onUserInteraction();
 	}
 
-    @Override
-    public void onPause() {
-        super.onPause();
-    	passcodeManager.onPause(this);
-    }
+	@Override
+	public void onPause() {
+		super.onPause();
+		delegate.onPause();
+	}
 
-    @Override
-    public void onDestroy() {
-    	unregisterReceiver(userSwitchReceiver);
-        unregisterReceiver(logoutCompleteReceiver);
-    	super.onDestroy();
-    }
+	@Override
+	public void onDestroy() {
+		delegate.onDestroy();
+		super.onDestroy();
+	}
 
 	@Override
 	public boolean onKeyUp(int keyCode, KeyEvent event) {
-		if (SalesforceSDKManager.getInstance().isDevSupportEnabled()) {
-			if (keyCode == KeyEvent.KEYCODE_MENU) {
-				SalesforceSDKManager.getInstance().showDevSupportDialog(this);
-				return true;
-			}
-		}
-		return false;
+		return delegate.onKeyUp(keyCode, event);
 	}
 
-    /**
-     * Refreshes the client if the user has been switched.
-     */
-	protected void refreshIfUserSwitched() {
-		if (passcodeManager.onResume(this)) {
-
-			// Gets login options.
-			final String accountType = SalesforceSDKManager.getInstance().getAccountType();
-	    	final LoginOptions loginOptions = SalesforceSDKManager.getInstance().getLoginOptions();
-
-			// Gets a rest client.
-			new ClientManager(this, accountType, loginOptions,
-					SalesforceSDKManager.getInstance().shouldLogoutWhenTokenRevoked()).getRestClient(this, new RestClientCallback() {
-
-				@Override
-				public void authenticatedRestClient(RestClient client) {
-					if (client == null) {
-						SalesforceSDKManager.getInstance().logout(SalesforceListActivity.this);
-						return;
-					}
-					onResume(client);
-
-					// Lets observers know that rendition is complete.
-					EventsObservable.get().notifyEvent(EventType.RenditionComplete);
-				}
-			});
-		}
-	}
-
-    /**
-     * Performs actions on logout complete.
-     */
-    protected void logoutCompleteActions() {
-    }
-
-    /**
-     * Acts on the user switch event.
-     *
-     * @author bhariharan
-     */
-    private class ActivityUserSwitchReceiver extends UserSwitchReceiver {
-
-		@Override
-		protected void onUserSwitch() {
-			refreshIfUserSwitched();
-		}
-    }
-
-	/**
-	 * Acts on the logout complete event.
-	 *
-	 * @author bhariharan
-	 */
-	private class ActivityLogoutCompleteReceiver extends LogoutCompleteReceiver {
-
-		@Override
-		protected void onLogoutComplete() {
-            logoutCompleteActions();
-		}
+	@Override
+	public void logoutCompleteActions() {
 	}
 }
